增加Spring框架依賴，方便看源碼

+++ src/test/java/fly/leetcode/cn/q82/SolutionTest.java
@@ -31,7 +31,7 @@ class SolutionTest {
 
         int i = 0;
         while (output!= null){
-            assertEquals(res.getInteger(i++),output.val);
+            assertEquals(res.getInteger(i++).intValue(),output.val);
             output = output.next;
         }
         assertEquals(i,res.size());
